Updated redme.md
An application developed for the maintenance of a nursing home. It is possible to keep track of the list of protégés and information about them. List all the personal belongings of each resident and their findings, a list of their staff and the user of the application. The application has different levels of use. It is possible to keep track of users of rooms and information about all rooms and users.